Critical section template

+++ BouncingBalls/Logic/RandomGenerator.cs
@@ -9,9 +9,9 @@
             _random = new Random();
         }
         
-        public double GenerateDouble(double min, double max)
+        public float GenerateDouble(double min, double max)
         {
-            return _random.NextDouble() * (max - min) + min;
+            return (float) (_random.NextDouble() * (max - min) + min);
         }
 
         public double GenerateDouble()
